Fix macros for Scala 2.10
