+++ monix-execution/shared/src/main/scala/monix/execution/Ack.scala
@@ -17,12 +17,10 @@
 
 package monix.execution
 
-import monix.execution.misc.{HygieneUtilMacros, InlineMacros}
 import scala.concurrent.duration.Duration
-import scala.concurrent.{Promise, CanAwait, ExecutionContext, Future}
-import scala.reflect.macros.whitebox
-import scala.util.{Failure, Success, Try}
+import scala.concurrent.{CanAwait, ExecutionContext, Future, Promise}
 import scala.language.experimental.macros
+import scala.util.{Failure, Success, Try}
 
 /** Represents an acknowledgement of processing that a consumer
   * sends back upstream. Useful to implement back-pressure.
@@ -177,287 +175,6 @@ object Ack {
         else
           source
       }
-  }
-
-  /** Macro implementations for [[AckExtensions]]. */
-  @macrocompat.bundle
-  class Macros(override val c: whitebox.Context) extends InlineMacros with HygieneUtilMacros {
-    import c.universe._
-
-    def isSynchronous[Self <: Future[Ack] : c.WeakTypeTag]: c.Expr[Boolean] = {
-      val selfExpr = sourceFrom[Self](c.prefix.tree)
-      val self = util.name("source")
-      val ContinueSymbol = symbolOf[Continue].companion
-      val StopSymbol = symbolOf[Stop].companion
-
-      val tree =
-        if (util.isClean(selfExpr))
-          q"""($selfExpr eq $ContinueSymbol) || ($selfExpr eq $StopSymbol)"""
-        else
-          q"""
-          val $self = $selfExpr
-          ($self eq $ContinueSymbol) || ($self eq $StopSymbol)
-          """
-
-      inlineAndReset[Boolean](tree)
-    }
-
-    def syncOnContinue[Self <: Future[Ack] : c.WeakTypeTag](callback: Tree)(s: Tree): Tree = {
-      val selfExpr = sourceFrom[Self](c.prefix.tree)
-      val self = util.name("source")
-      val scheduler = c.Expr[Scheduler](s)
-
-      val execute = c.Expr[Unit](callback)
-      val ContinueSymbol = symbolOf[Continue].companion
-      val StopSymbol = symbolOf[Stop].companion
-      val AckSymbol = symbolOf[Ack]
-      val FutureSymbol = symbolOf[Future[_]]
-
-      val tree =
-        q"""
-        val $self = $selfExpr
-        if ($self eq $ContinueSymbol)
-          try { $execute } catch {
-            case ex: Throwable =>
-              if (_root_.scala.util.control.NonFatal(ex))
-                $scheduler.reportFailure(ex)
-              else
-                throw ex
-          }
-        else if (($self : $FutureSymbol[$AckSymbol]) != $StopSymbol) {
-          $self.onComplete { result =>
-            if (result.isSuccess && (result.get eq $ContinueSymbol)) { $execute }
-          }($scheduler)
-        }
-
-        $self
-        """
-
-      inlineAndResetTree(tree)
-    }
-
-    def syncOnStopOrFailure[Self <: Future[Ack] : c.WeakTypeTag](callback: Tree)(s: Tree): Tree = {
-      val selfExpr = sourceFrom[Self](c.prefix.tree)
-      val self = util.name("source")
-      val scheduler = c.Expr[Scheduler](s)
-
-      val execute = c.Expr[Unit](callback)
-      val ContinueSymbol = symbolOf[Continue].companion
-      val StopSymbol = symbolOf[Stop].companion
-      val AckSymbol = symbolOf[Ack]
-      val FutureSymbol = symbolOf[Future[_]]
-
-      val tree =
-        q"""
-        val $self = $selfExpr
-        if ($self eq $StopSymbol)
-          try { $execute } catch {
-            case ex: Throwable =>
-              if (_root_.scala.util.control.NonFatal(ex))
-                $scheduler.reportFailure(ex)
-              else
-                throw ex
-          }
-        else if (($self : $FutureSymbol[$AckSymbol]) != $ContinueSymbol) {
-          $self.onComplete { result =>
-            if (result.isFailure || (result.get eq $StopSymbol)) { $execute }
-          }($scheduler)
-        }
-
-        $self
-        """
-
-      inlineAndResetTree(tree)
-    }
-
-    def syncMap[Self <: Future[Ack] : c.WeakTypeTag](f: c.Expr[Ack => Ack])(s: c.Expr[Scheduler]): c.Expr[Future[Ack]] = {
-      val selfExpr = sourceFrom[Self](c.prefix.tree)
-      val schedulerExpr = s
-      val self = util.name("source")
-      val fn = util.name("fn")
-
-      val ContinueSymbol = symbolOf[Continue].companion
-      val StopSymbol = symbolOf[Stop].companion
-      val AckSymbol = symbolOf[Ack]
-
-      val tree =
-        if (util.isClean(f)) {
-          q"""
-          val $self = $selfExpr
-
-          if (($self eq $ContinueSymbol) || ($self eq $StopSymbol)) {
-            try {
-              $f($self.asInstanceOf[$AckSymbol]) : $AckSymbol
-            } catch {
-              case ex: _root_.java.lang.Throwable =>
-                if (_root_.scala.util.control.NonFatal(ex)) {
-                  $schedulerExpr.reportFailure(ex)
-                  $StopSymbol
-                } else {
-                  throw ex
-                }
-            }
-          } else {
-            $self.map($f)
-          }
-          """
-        } else {
-          q"""
-          val $self = $selfExpr
-          val $fn: _root_.scala.Function1[$AckSymbol,$AckSymbol] = $f
-
-          if (($self eq $ContinueSymbol) || ($self eq $StopSymbol))
-            try {
-              $fn($self.asInstanceOf[$AckSymbol]) : $AckSymbol
-            } catch {
-              case ex: Throwable =>
-                if (_root_.scala.util.control.NonFatal(ex)) {
-                  $schedulerExpr.reportFailure(ex)
-                  $StopSymbol
-                } else {
-                  throw ex
-                }
-            }
-          else {
-            $self.map($fn)
-          }
-          """
-        }
-
-      inlineAndReset[Future[Ack]](tree)
-    }
-
-    def syncFlatMap[Self <: Future[Ack] : c.WeakTypeTag](f: c.Expr[Ack => Future[Ack]])(s: c.Expr[Scheduler]): c.Expr[Future[Ack]] = {
-      val selfExpr = sourceFrom[Self](c.prefix.tree)
-      val schedulerExpr = s
-      val self = util.name("source")
-
-      val ContinueSymbol = symbolOf[Continue].companion
-      val StopSymbol = symbolOf[Stop].companion
-      val AckSymbol = symbolOf[Ack]
-      val FutureSymbol = symbolOf[Future[_]]
-
-      val tree =
-        if (util.isClean(f))
-          q"""
-          val $self = $selfExpr
-
-          if (($self eq $ContinueSymbol) || ($self eq $StopSymbol))
-            try {
-              $f($self.asInstanceOf[$AckSymbol]) : $FutureSymbol[$AckSymbol]
-            } catch {
-              case ex: Throwable =>
-                if (_root_.scala.util.control.NonFatal(ex)) {
-                  $schedulerExpr.reportFailure(ex)
-                  $StopSymbol
-                } else {
-                  throw ex
-                }
-            }
-          else {
-            $self.flatMap($f)
-          }
-          """
-        else {
-          val fn = util.name("fn")
-          q"""
-          val $self = $selfExpr
-          val $fn: _root_.scala.Function1[$AckSymbol,$AckSymbol] = $f
-
-          if (($self eq $ContinueSymbol) || ($self eq $StopSymbol))
-            try {
-              $fn($self.asInstanceOf[$AckSymbol]) : $FutureSymbol[$AckSymbol]
-            } catch {
-              case ex: Throwable =>
-                if (_root_.scala.util.control.NonFatal(ex)) {
-                  $schedulerExpr.reportFailure(ex)
-                  $StopSymbol
-                } else {
-                  throw ex
-                }
-            }
-          else {
-            $self.flatMap($fn)
-          }
-          """
-        }
-
-      inlineAndReset[Future[Ack]](tree)
-    }
-
-    def syncOnComplete[Self <: Future[Ack] : c.WeakTypeTag](f: c.Expr[scala.util.Try[Ack] => Unit])
-      (s: c.Expr[Scheduler]): c.Expr[Unit] = {
-
-      val selfExpr = sourceFrom[Self](c.prefix.tree)
-      val schedulerExpr = s
-      val self = util.name("source")
-
-      val SuccessObject = symbolOf[scala.util.Success[_]].companion
-      val ContinueObject = symbolOf[Continue].companion
-      val StopObject = symbolOf[Stop].companion
-      val AckSymbol = symbolOf[Ack]
-      val TrySymbol = symbolOf[scala.util.Try[_]]
-      val UnitSymbol = symbolOf[Unit]
-      val FutureSymbol = symbolOf[Future[_]]
-
-      val tree =
-        if (util.isClean(f))
-          q"""
-          val $self: $FutureSymbol[$AckSymbol] = $selfExpr
-
-          if (($self eq $ContinueObject) || ($self eq $StopObject))
-            try {
-              $f($SuccessObject($self.asInstanceOf[$AckSymbol]) : $TrySymbol[$AckSymbol])
-              ()
-            } catch {
-              case ex: Throwable =>
-                if (_root_.scala.util.control.NonFatal(ex)) {
-                  $schedulerExpr.reportFailure(ex)
-                } else {
-                  throw ex
-                }
-            }
-          else {
-            $self.onComplete($f)
-          }
-          """
-        else {
-          val fn = util.name("fn")
-          q"""
-          val $self: $FutureSymbol[$AckSymbol]  = $selfExpr
-          val $fn: _root_.scala.Function1[$TrySymbol[$AckSymbol],$UnitSymbol] = $f
-
-          if (($self eq $ContinueObject) || ($self eq $StopObject))
-            try {
-              $fn($SuccessObject($self.asInstanceOf[$AckSymbol]))
-              ()
-            } catch {
-              case ex: Throwable =>
-                if (_root_.scala.util.control.NonFatal(ex)) {
-                  $schedulerExpr.reportFailure(ex)
-                } else {
-                  throw ex
-                }
-            }
-          else {
-            $self.onComplete($fn)
-          }
-          """
-        }
-
-      inlineAndReset[Unit](tree)
-    }
-
-    private[monix] def sourceFrom[Source : c.WeakTypeTag](tree: Tree): c.Expr[Source] = {
-      val ackExtensions = symbolOf[AckExtensions[_]].name.toTermName
-      tree match {
-        case Apply(TypeApply(Select(_, `ackExtensions`), _), List(expr)) =>
-          c.Expr[Source](expr)
-        case _ =>
-          c.warning(tree.pos, "Could not infer the implicit class source, please report a bug!")
-          c.Expr[Source](q"$tree.source")
-      }
-    }
   }
 }
